Fix: textarea height not adjust when submit a message
Due to: react-hook-form's `reset` function, when called it, this is not gonna trigger an `onChange` event.
Solve by: using reac-hook-form's `watch` function. This meant for monitor a change value of the target element.
On `useAdjustTextareaHeight` add parameter to receive a dependency for component to know when it should be adjusted based on which value.